renamed ThreadFactory.javaFactory
