+++ src/main/java/executor/service/service/parallel/ExecutionWorker.java
@@ -2,6 +2,7 @@ package executor.service.service.parallel;
 
 import executor.service.model.ProxyConfigHolder;
 import executor.service.model.Scenario;
+import executor.service.service.ExecutionService;
 
 /**
  * Execution tasks.
